+++ studies/florida_straits/flstrts_slope_test.cc
@@ -15,7 +15,7 @@ int main( int argc, char* argv[] ) {
 
     ascii_arc_bathy* bathymetry = new ascii_arc_bathy( 
 	USML_STUDIES_DIR "/florida_straits/flstrts_bathymetry.asc" ) ;
-    boundary_grid<float,2> bottom( bathymetry ) ;
+    boundary_grid<double,2> bottom( bathymetry ) ;
 
     std::ofstream file1( USML_STUDIES_DIR "/florida_straits/flstrts_slope_depth.csv" ) ;
     std::ofstream file2( USML_STUDIES_DIR "/florida_straits/flstrts_slope_normlat.csv" ) ;
