+++ waveq3d/test/reflection_test.cc
@@ -371,7 +371,7 @@ BOOST_AUTO_TEST_CASE( reflect_grid_test ) {
         // load bathymetry from ETOPO1 database
 
         cout << "load bathymetry" << endl ;
-        boundary_model* bottom = new boundary_grid<float,2>( new netcdf_bathy(
+        boundary_model* bottom = new boundary_grid<double,2>( new netcdf_bathy(
     	    USML_DATA_DIR "/bathymetry/ETOPO1_Ice_g_gmt4.grd",
             lat1, lat2, lng1, lng2 ) ) ;
